AEA-1349: (Some of the) review changes. Fix sandbox pact tests.
