+++ models/fhir/common.ts
@@ -1,7 +1,4 @@
 import {LosslessNumber} from "lossless-json"
-import { PersonOrOrganization } from "."
-import { fhir } from ".."
-import { Organization } from "../hl7-v3"
 import {Extension} from "./extension"
 
 export abstract class Element {
@@ -54,7 +51,6 @@ export function createCoding(system: string, code: string, display: string): Cod
   }
 }
 
-// Matt and Dan -
 export interface Reference<T extends Resource> {
   reference: string,
   display?: string
