+++ tests/e2e/pact/specs/sandbox/process.successes.spec.ts
@@ -182,60 +182,6 @@ jestpact.pactWith(
   })
 
 jestpact.pactWith(
-  pactOptions("sandbox", "process", "claim"),
-  /* eslint-disable  @typescript-eslint/no-explicit-any */
-  async (provider: any) => {
-    const client = () => {
-      const url = `${provider.mockService.baseUrl}`
-      return supertest(url)
-    }
-
-    describe("process-message claim sandbox e2e tests", () => {
-      test.each(TestResources.processClaimInformationCases)(
-        "should be able to claim %s",
-        async (desc: string, message: fhir.Bundle) => {
-          const apiPath = `${basePath}/$process-message`
-          const bundleStr = LosslessJson.stringify(message)
-          const bundle = JSON.parse(bundleStr) as fhir.Bundle
-
-          const requestId = uuid.v4()
-          const correlationId = uuid.v4()
-
-          const interaction: InteractionObject = {
-            state: "is authenticated",
-            uponReceiving: `a request to process ${desc} message to Spine`,
-            withRequest: {
-              headers: {
-                "Content-Type": "application/fhir+json; fhirVersion=4.0",
-                "X-Request-ID": requestId,
-                "X-Correlation-ID": correlationId
-              },
-              method: "POST",
-              path: apiPath,
-              body: bundle
-            },
-            willRespondWith: {
-              headers: {
-                "Content-Type": "application/json"
-              },
-              //TODO - Verify response body for claims
-              status: 200
-            }
-          }
-          await provider.addInteraction(interaction)
-          await client()
-            .post(apiPath)
-            .set("Content-Type", "application/fhir+json; fhirVersion=4.0")
-            .set("X-Request-ID", requestId)
-            .set("X-Correlation-ID", correlationId)
-            .send(bundleStr)
-            .expect(200)
-        }
-      )
-    })
-  })
-
-jestpact.pactWith(
   pactOptions("sandbox", "process", "send"),
   /* eslint-disable  @typescript-eslint/no-explicit-any */
   async (provider: any) => {
